REFACTOR: Add components directory
